+++ tests/test_unit_errors.py
@@ -6,7 +6,8 @@ from unittest.mock import MagicMock
 
 from test_unit_common import UnitTest
 
-from arduino.router_bridge import BUFFER_LIMIT_EXCEEDED_ERR, GENERIC_ERR, ClientServer
+from arduino.router_bridge.bridge import ClientServer
+from arduino.router_bridge.connection import BUFFER_LIMIT_EXCEEDED_ERR, GENERIC_ERR
 
 
 class TestErrors(UnitTest):